requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cryptography==43.0.3
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dydantic==0.0.7
executing==2.1.0
fastjsonschema==2.21.1
fireworks-ai==0.15.11
flake8==7.1.2
Flask==3.1.0
fqdn==1.5.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema_rs==0.25.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langchain==0.3.21
langchain-anthropic==0.3.1
langchain-community==0.3.20
langchain-core==0.3.47
langchain-fireworks==0.2.6
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langgraph==0.3.18
langgraph-api==0.0.15
langgraph-checkpoint==2.0.21
langgraph-checkpoint-sqlite==2.0.1
langgraph-cli==0.1.65
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.48
langsmith==0.2.10
MarkupSafe==3.0.2
marshmallow==3.23.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
nulltype==2.3.1
numpy==1.26.4
openai==1.68.2
orjson==3.10.13
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
plaid-python==29.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
PyYAML==6.0.2
pyzmq==26.2.0
-e git+https://github.com/langchain-ai/langchain-academy.git@37168db0825602f982aac19aa6b34f920e2c399c#egg=react_agent&subdirectory=langgraph_studio
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.45.2
structlog==24.4.0
tavily-python==0.5.1
tenacity==8.5.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trustcall==0.0.26
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.18.3
